Refactor GeoMapViewer Point hit detection
